Randomizes the board created
Created a method in Board.py that randomizes the board
that is created. This will allow for a more random gameplay.
Though there is an issue where sometimes the program will
reach the maximum recursion rate when solving for a specific
Sudoku board. I do not think there is any way to solve this issue.

Resolves: #12